slate: update layouts for retina display
The retina display on my new laptop has enough resolution that I can
actually do reasonable tiling on that display alone, so it's time to
dust off the layouts.

I've got my 2-screen layout as 0, a fully tiled layout as 1 and two
different overlapped tilings as 2 and 3.

Also fix up some of the other configuration directives that weren't
working.