[zshrc] update zshrc for new etsy laptop
Some things have changed. My hostname is the inventory ID number,
homebrew is a bit different than it used to be, and the google cloud sdk
is in a different spot.

I also finally added a line to source the cloudsnooze files so I can
control my vm without jumping through hoops